+++ bittytax/conv/parsers/binance.py
@@ -15,9 +15,10 @@ from ..exceptions import UnexpectedTypeError, UnexpectedTradingPairError, \
 
 WALLET = "Binance"
 
-QUOTE_ASSETS = ['AUD', 'BIDR', 'BKRW', 'BNB', 'BRL', 'BTC', 'BUSD', 'BVND', 'DAI', 'ETH',
-                'EUR', 'GBP', 'GYEN', 'IDRT', 'NGN', 'PAX', 'RUB', 'TRX', 'TRY', 'TUSD',
-                'UAH', 'USDC', 'USDP', 'USDS', 'USDT', 'VAI', 'XRP', 'ZAR']
+QUOTE_ASSETS = ['AUD', 'BIDR', 'BKRW', 'BNB', 'BRL', 'BTC', 'BUSD', 'BVND', 'DAI', 'DOGE',
+                'DOT', 'ETH', 'EUR', 'GBP', 'GYEN', 'IDRT', 'NGN', 'PAX', 'RUB', 'TRX',
+                'TRY', 'TUSD', 'UAH', 'USDC', 'USDP', 'USDS', 'USDT', 'UST', 'VAI', 'XRP',
+                'ZAR']
 
 BASE_ASSETS = ['1INCH', '1INCHDOWN', '1INCHUP']
 
@@ -52,6 +53,26 @@ def parse_binance_trades(data_row, parser, **_kwargs):
                                                  wallet=WALLET)
     else:
         raise UnexpectedTypeError(parser.in_header.index('Type'), 'Type', row_dict['Type'])
+
+def parse_binance_convert(data_row, parser, **_kwargs):
+    row_dict = data_row.row_dict
+    data_row.timestamp = DataParser.parse_timestamp(row_dict['Date'])
+
+    if row_dict['Status'] != "Successful":
+        return
+
+    base_asset, quote_asset = split_trading_pair(row_dict['Pair'])
+    if base_asset is None or quote_asset is None:
+        raise UnexpectedTradingPairError(parser.in_header.index('Pair'), 'Pair',
+                                         row_dict['Pair'])
+
+    data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
+                                             data_row.timestamp,
+                                             buy_quantity=row_dict['Buy'].split(' ')[0],
+                                             buy_asset=row_dict['Buy'].split(' ')[1],
+                                             sell_quantity=row_dict['Sell'].split(' ')[0],
+                                             sell_asset=row_dict['Sell'].split(' ')[1],
+                                             wallet=WALLET)
 
 def parse_binance_trades_statement(data_row, parser, **_kwargs):
     row_dict = data_row.row_dict
@@ -172,49 +193,59 @@ def parse_binance_statements(data_rows, parser, **_kwargs):
         row_dict = data_row.row_dict
         data_row.timestamp = DataParser.parse_timestamp(row_dict['UTC_Time'])
 
-        if row_dict['Operation'] in ("Commission History", "Referrer rebates"):
+        if row_dict['Operation'] in ("Commission History", "Referrer rebates", "Commission Rebate",
+                                     "Commission Fee Shared With You", "Cash Voucher distribution",
+                                     "Referral Kickback"):
             data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_GIFT_RECEIVED,
                                                      data_row.timestamp,
                                                      buy_quantity=row_dict['Change'],
                                                      buy_asset=row_dict['Coin'],
                                                      wallet=WALLET)
         elif row_dict['Operation'] == "Distribution":
-            data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_AIRDROP,
-                                                     data_row.timestamp,
-                                                     buy_quantity=row_dict['Change'],
-                                                     buy_asset=row_dict['Coin'],
-                                                     wallet=WALLET)
+            if Decimal(row_dict['Change']) > 0:
+                data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_AIRDROP,
+                                                         data_row.timestamp,
+                                                         buy_quantity=row_dict['Change'],
+                                                         buy_asset=row_dict['Coin'],
+                                                         wallet=WALLET)
+            else:
+                data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_SPEND,
+                                                         data_row.timestamp,
+                                                         sell_quantity= \
+                                                             abs(Decimal(row_dict['Change'])),
+                                                         sell_asset=row_dict['Coin'],
+                                                         wallet=WALLET)
         elif row_dict['Operation'] == "Super BNB Mining":
             data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_MINING,
                                                      data_row.timestamp,
                                                      buy_quantity=row_dict['Change'],
                                                      buy_asset=row_dict['Coin'],
                                                      wallet=WALLET)
-        elif row_dict['Operation'] == "Savings Interest":
+        elif row_dict['Operation'] in ("Savings Interest", "Pool Distribution"):
             data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_INTEREST,
                                                      data_row.timestamp,
                                                      buy_quantity=row_dict['Change'],
                                                      buy_asset=row_dict['Coin'],
                                                      wallet=WALLET)
-        elif row_dict['Operation'] == "POS savings interest":
+        elif row_dict['Operation'] in ("POS savings interest", "ETH 2.0 Staking Rewards",
+                                       "Liquid Swap rewards"):
             data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_STAKING,
                                                      data_row.timestamp,
                                                      buy_quantity=row_dict['Change'],
                                                      buy_asset=row_dict['Coin'],
                                                      wallet=WALLET)
-        elif row_dict['Operation'] == "Small assets exchange BNB":
-            bnb_rows = [dr for dr in tx_times[row_dict['UTC_Time']]
-                        if dr.row_dict['Operation'] == row_dict['Operation']]
-            bnb_convert(bnb_rows)
+        elif row_dict['Operation'] in ("Small assets exchange BNB", "Large OTC trading"):
+            make_trade(row_dict['Operation'], tx_times[row_dict['UTC_Time']])
         elif row_dict['Operation'] in ("Savings purchase", "Savings Principal redemption",
                                        "POS savings purchase", "POS savings redemption"):
             # Skip not taxable events
             continue
 
-def bnb_convert(bnb_rows):
-    buy_quantity = get_bnb_quantity(bnb_rows)
+def make_trade(operation, tx_times):
+    op_rows = [dr for dr in tx_times if dr.row_dict['Operation'] == operation]
+    buy_quantity, buy_asset = get_buy_quantity(op_rows)
 
-    for data_row in bnb_rows:
+    for data_row in op_rows:
         if not data_row.parsed:
             data_row.timestamp = DataParser.parse_timestamp(data_row.row_dict['UTC_Time'])
             data_row.parsed = True
@@ -222,42 +253,60 @@ def bnb_convert(bnb_rows):
             data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
                                                      data_row.timestamp,
                                                      buy_quantity=buy_quantity,
-                                                     buy_asset="BNB",
+                                                     buy_asset=buy_asset,
                                                      sell_quantity=abs(Decimal(data_row. \
                                                          row_dict['Change'])),
                                                      sell_asset=data_row.row_dict['Coin'],
                                                      wallet=WALLET)
 
-def get_bnb_quantity(bnb_rows):
-    bnb_found = False
+def get_buy_quantity(op_rows):
+    buy_found = False
     buy_quantity = None
-    assets = 0
+    buy_asset = ''
+    sell_assets = 0
 
-    for data_row in bnb_rows:
-        if data_row.row_dict['Coin'] == "BNB":
+    for data_row in op_rows:
+        if Decimal(data_row.row_dict['Change']) > 0:
             data_row.timestamp = DataParser.parse_timestamp(data_row.row_dict['UTC_Time'])
             data_row.parsed = True
 
-            if not bnb_found:
+            if not buy_found:
                 buy_quantity = data_row.row_dict['Change']
-                bnb_found = True
+                buy_asset = data_row.row_dict['Coin']
+                buy_found = True
             else:
-                # Multiple BNB quantities, will need to be added manually
+                # Multiple buys, quantity will need to be added manually
                 buy_quantity = None
+                buy_asset = ''
         else:
-            assets += 1
+            sell_assets += 1
 
-    if assets > 1:
-        # Multiple assets converted, BNB quantities will need to be added manually
+    if sell_assets > 1:
+        # Multiple sells, quantity will need to be added manually
         buy_quantity = None
+        buy_asset = ''
 
-    return buy_quantity
+    return buy_quantity, buy_asset
 
 DataParser(DataParser.TYPE_EXCHANGE,
            "Binance Trades",
            ['Date(UTC)', 'Market', 'Type', 'Price', 'Amount', 'Total', 'Fee', 'Fee Coin'],
            worksheet_name="Binance T",
            row_handler=parse_binance_trades)
+
+DataParser(DataParser.TYPE_EXCHANGE,
+           "Binance Trades",
+           ['Date', 'Pair', 'Type', 'Sell', 'Buy', 'Price', 'Inverse Price', 'Date Updated',
+            'Status'],
+           worksheet_name="Binance T",
+           row_handler=parse_binance_convert)
+
+DataParser(DataParser.TYPE_EXCHANGE,
+           "Binance Trades",
+           ['Date', 'Wallet', 'Pair', 'Type', 'Sell', 'Buy', 'Price', 'Inverse Price',
+            'Date Updated', 'Status'],
+           worksheet_name="Binance T",
+           row_handler=parse_binance_convert)
 
 DataParser(DataParser.TYPE_EXCHANGE,
            "Binance Trades",